+++ apps/imlib/src/imboy_cipher.erl
@@ -42,11 +42,9 @@ aes_decrypt(Type, Bin, Key, IV) when is_binary(Key) ->
     aes_decrypt(Type, Bin, binary_to_list(Key), IV);
 aes_decrypt(Type, Bin, Key, IV) ->
     Bin1 = base64:decode(Bin),
-    IV = config_ds:env(solidified_key_iv),
     StateDec = crypto:crypto_init(Type, Key, IV, false),
     Bin2 = crypto_update(StateDec, Bin1, size(Bin1), <<>>),
     binary:part(Bin2, {0, size(Bin2) - binary:last(Bin2)}).
-
 
 -spec rsa_encrypt(CipherText :: binary(), PrivKey :: binary()) -> PlainText :: binary().