+++ doc/postgresql/migrations/00000043_imboy_v0.4.0.sql
@@ -1,0 +1,4 @@
+
+ALTER TABLE public.user_device ADD COLUMN IF NOT EXISTS public_key VARCHAR(1024) DEFAULT '';
+
+COMMENT ON COLUMN public.user_device.public_key IS '用户设备公钥';